requirements.txt
opencv-python
pytesseract
pandas
Pillow
streamlit
pdf2image
openpyxl


